fix(consent): keep malformed-token error opaque, reject regressive cache eviction (CWE-209)
validate_token() now returns a generic "Malformed token" reason instead of
leaking str(e) details for unexpected parse failures, matching the existing
opaque-error pattern already used elsewhere in the function.

The size-cap eviction behavior proposed earlier in this branch is not
included: evicting the soonest-expiring revocation entry under cache
pressure would let a revoked-but-unexpired token validate as not-revoked
again, contradicting the cache's documented design intent and the existing
test_size_cap_does_not_evict_unexpired_revocations test. Renamed the test
file to reflect that it now only covers the opaque-message fix.
