+++ consent-protocol/hushh_mcp/consent/token.py
@@ -45,15 +45,8 @@ class _BoundedRevocationCache:
         with self._lock:
             self._evict_expired_locked(now_ms)
             if len(self._entries) >= self._MAX_SIZE:
-                # TTL eviction alone did not free space — all remaining entries
-                # are still within their grace window. Evict the one whose
-                # revocation marker will become redundant soonest (smallest
-                # evict_after_ms), because once a token's own expiry passes,
-                # validate_token() rejects it independently of the cache.
-                oldest_key = min(self._entries, key=lambda k: self._entries[k])
-                del self._entries[oldest_key]
                 logger.warning(
-                    "revocation_cache.size_cap_hit evicted_oldest size=%s max_size=%s",
+                    "revocation_cache.size_cap_exceeded size=%s max_size=%s",
                     len(self._entries),
                     self._MAX_SIZE,
                 )
